+++ OpenSim/Common/SmoothSegmentedFunction.cpp
@@ -26,6 +26,11 @@
 #include "SmoothSegmentedFunction.h"
 #include <fstream>
 #include "simmath/internal/SplineFitter.h"
+#include <memory>
+#include <mutex>
+#include <unordered_map>
+#include <utility>
+#include <cmath>
 
 //=============================================================================
 // STATICS
@@ -39,6 +44,312 @@ static double UTOL = (double)SimTK::Eps*1e2;
 static double INTTOL = (double)SimTK::Eps*1e2;
 static int MAXITER = 20;
 static int NUM_SAMPLE_PTS = 100;
+
+//=============================================================================
+// PARAMETERS
+//=============================================================================
+
+// Helper struct containing all parameters required to construct the
+// SmoothSegmentedFunctionData.
+namespace {
+struct SmoothSegmentedFunctionParameters
+{
+
+    SmoothSegmentedFunctionParameters(
+        const SimTK::Matrix& mX,
+        const SimTK::Matrix& mY,
+        double x0,
+        double x1,
+        double y0,
+        double y1,
+        double dydx0,
+        double dydx1,
+        bool computeIntegral,
+        bool intx0x1) :
+        _mX(mX),
+        _mY(mY),
+        _x0(x0),
+        _x1(x1),
+        _y0(y0),
+        _y1(y1),
+        _dydx0(dydx0),
+        _dydx1(dydx1),
+        _computeIntegral(computeIntegral),
+        _intx0x1(intx0x1) {}
+
+    SmoothSegmentedFunctionParameters() {}
+
+    SimTK::Matrix _mX = SimTK::Matrix();
+    SimTK::Matrix _mY = SimTK::Matrix();
+    double _x0 = SimTK::NaN;
+    double _x1 = SimTK::NaN;
+    double _y0 = SimTK::NaN;
+    double _y1 = SimTK::NaN;
+    double _dydx0 = SimTK::NaN;
+    double _dydx1 = SimTK::NaN;
+    bool _computeIntegral = false;
+    bool _intx0x1 = false;
+};
+
+bool operator==(
+    const SimTK::Matrix& lhs,
+    const SimTK::Matrix& rhs)
+{
+    if (lhs.nrow() != rhs.nrow() || lhs.ncol() != rhs.ncol()) {
+        return false;
+    }
+    for (int r = 0; r < lhs.nrow(); ++r) {
+        for (int c = 0; c < lhs.ncol(); ++c) {
+            if (lhs.row(r)[c] != rhs.row(r)[c]) {
+                return false;
+            }
+        }
+    }
+    return true;
+}
+
+// Does what you expect, with the exception of NaN==NaN resulting in true, for
+// x0, x1, y0, y1, dydxo, dydx1. This is to catch the uninitialized case.
+bool operator==(
+    const SmoothSegmentedFunctionParameters& lhs,
+    const SmoothSegmentedFunctionParameters& rhs)
+{
+    auto equalOrBothNaN = [] (double a, double b) -> bool {
+        return a == b || ( std::isnan(a) && std::isnan(b) );
+    };
+    return
+        lhs._mX == rhs._mX &&
+        lhs._mY == rhs._mY &&
+        equalOrBothNaN(lhs._x0, rhs._x0) &&
+        equalOrBothNaN(lhs._x1, rhs._x1) &&
+        equalOrBothNaN(lhs._y0, rhs._y0) &&
+        equalOrBothNaN(lhs._y1, rhs._y1) &&
+        equalOrBothNaN(lhs._dydx0, rhs._dydx0) &&
+        equalOrBothNaN(lhs._dydx1, rhs._dydx1) &&
+        lhs._computeIntegral == rhs._computeIntegral &&
+        lhs._intx0x1 == rhs._intx0x1;
+}
+
+} // namespace
+
+//=============================================================================
+// HASHING OF PARAMETERS
+//=============================================================================
+
+template <typename T>
+static inline size_t HashCombine(size_t seed, const T& v)
+{
+    std::hash<T> hasher;
+    return seed ^ (hasher(v) + 0x9e3779b9 + (seed << 6) + (seed >> 2));
+}
+
+template <typename T>
+static inline size_t HashOf(const T& v)
+{
+    return std::hash<T>{}(v);
+}
+
+template <typename T, typename... Others>
+static inline size_t HashOf(const T& v, const Others&... others)
+{
+    return HashCombine(HashOf(v), HashOf(others...));
+}
+
+template <>
+inline size_t HashOf(const SimTK::Matrix& matrix)
+{
+    size_t hash = HashOf(matrix.nrow(), matrix.ncol());
+    for (int r = 0; r < matrix.nrow(); ++r) {
+        for (int c = 0; c < matrix.ncol(); ++c) {
+            hash = HashCombine(hash, matrix.row(r)[c]);
+        }
+    }
+    return hash;
+}
+
+template <> struct std::hash<SmoothSegmentedFunctionParameters> final
+{
+    size_t operator()(const SmoothSegmentedFunctionParameters& params) const
+    {
+        return HashOf(
+            params._mX,
+            params._mY,
+            params._x0,
+            params._x1,
+            params._y0,
+            params._y1,
+            params._dydx0,
+            params._dydx1,
+            params._computeIntegral,
+            params._intx0x1);
+    }
+};
+
+//=============================================================================
+// SMOOTHSEGMENTEDFUNCTION DATA
+//=============================================================================
+
+namespace OpenSim {
+struct SmoothSegmentedFunctionData
+{
+
+    SmoothSegmentedFunctionData(
+        const SimTK::Matrix& mX,
+        const SimTK::Matrix& mY,
+        double x0,
+        double x1,
+        double y0,
+        double y1,
+        double dydx0,
+        double dydx1,
+        bool computeIntegral,
+        bool intx0x1,
+        const std::string& name);
+
+    SmoothSegmentedFunctionData(
+        const SmoothSegmentedFunctionParameters& params,
+        const std::string& name) :
+        SmoothSegmentedFunctionData(
+            params._mX,
+            params._mY,
+            params._x0,
+            params._x1,
+            params._y0,
+            params._y1,
+            params._dydx0,
+            params._dydx1,
+            params._computeIntegral,
+            params._intx0x1,
+            name
+        ) {}
+
+    /**Array of spline fit functions X(u) for each Bezier elbow*/
+    SimTK::Array_<SimTK::Spline> _arraySplineUX;
+
+    /**Spline fit of the integral of the curve y(x)*/
+    SimTK::Spline _splineYintX;
+
+    /**Bezier X1,...,Xn control point locations. Control points are
+    stored in 6x1 vectors in the order above*/
+    SimTK::Array_<SimTK::Vector> _mXVec;
+
+    /**Bezier Y1,...,Yn control point locations. Control points are
+    stored in 6x1 vectors in the order above*/
+    SimTK::Array_<SimTK::Vector> _mYVec;
+
+    /**The number of quintic Bezier curves that describe the relation*/
+    int _numBezierSections;
+
+    /**The minimum value of the domain*/
+    double _x0;
+    /**The maximum value of the domain*/
+    double _x1;
+    /**The minimum value of the range*/
+    double _y0;
+    /**The maximum value of the range*/
+    double _y1;
+    /**The slope at _x0*/
+    double _dydx0;
+    /**The slope at _x1*/
+    double _dydx1;
+    /**This is the users */
+    bool _computeIntegral;
+
+    /**This variable, when true, indicates that the user wants the integral
+    from left to right (x0 to x1). If it is false, the integral from right
+    to left (x1 to x0) is computed*/
+    bool _intx0x1;
+};
+
+} // namespace OpenSim
+
+//=============================================================================
+// DATA LOOKUP
+//=============================================================================
+
+namespace {
+
+class SmoothSegmentedFunctionDataCache final
+{
+public:
+    std::shared_ptr<const SmoothSegmentedFunctionData> lookup(
+            const SmoothSegmentedFunctionParameters& params,
+            const std::string& name)
+    {
+        std::lock_guard<std::mutex> guard{_cacheMutex};
+        garbageCollectExpiredData();
+        return findOrInsert(params, name);
+    }
+
+private:
+    // Find previously constructed data, or construct new and insert.
+    std::shared_ptr<const SmoothSegmentedFunctionData> findOrInsert(
+            const SmoothSegmentedFunctionParameters& params,
+            const std::string& name)
+    {
+        auto it = _cache.find(params);
+        if (it != _cache.end()) {
+            auto data_ptr = it->second.lock();
+            // We expect expired data to be collected at this point, but check
+            // if expired here to be on the safe side.
+            if (data_ptr) {
+                return data_ptr;
+            }
+        }
+        return _cache.insert({
+                params,
+                std::make_shared<SmoothSegmentedFunctionData>(
+                    SmoothSegmentedFunctionData(params, name)
+                )
+            }).first->second.lock();
+    }
+
+    // Do a pass-over to clean up expired pointers.
+    void garbageCollectExpiredData()
+    {
+        for (auto it=_cache.begin(); it!=_cache.end();) {
+            if (it->second.expired()) {
+                it = _cache.erase(it);
+            } else {
+                ++it;
+            }
+        }
+    }
+
+    std::mutex _cacheMutex;
+    std::unordered_map<SmoothSegmentedFunctionParameters,
+        std::weak_ptr<const SmoothSegmentedFunctionData>> _cache;
+};
+
+std::shared_ptr<const OpenSim::SmoothSegmentedFunctionData>
+    SmoothSegmentedFunctionDataLookup(
+        const SmoothSegmentedFunctionParameters& params,
+        const std::string& name)
+{
+    static SmoothSegmentedFunctionDataCache s_GlobalCache;
+    return s_GlobalCache.lookup(params, name);
+}
+
+} // namespace
+
+//=============================================================================
+// RULE OF FIVE
+//=============================================================================
+
+SmoothSegmentedFunction::SmoothSegmentedFunction(
+        const SmoothSegmentedFunction&) = default;
+
+SmoothSegmentedFunction& SmoothSegmentedFunction::operator=(
+        const SmoothSegmentedFunction&) = default;
+
+SmoothSegmentedFunction::SmoothSegmentedFunction(
+        SmoothSegmentedFunction&&) noexcept = default;
+
+SmoothSegmentedFunction& SmoothSegmentedFunction::operator=(
+        SmoothSegmentedFunction&&) noexcept = default;
+
+SmoothSegmentedFunction::~SmoothSegmentedFunction() noexcept = default;
+
 //=============================================================================
 // UTILITY FUNCTIONS
 //=============================================================================
@@ -90,15 +401,27 @@ static int NUM_SAMPLE_PTS = 100;
 
         N.B. These costs are dependent on SegmentedQuinticBezierToolkit
 */
-SmoothSegmentedFunction::
-  SmoothSegmentedFunction(const SimTK::Matrix& mX, const SimTK::Matrix& mY,  
-          double x0, double x1, double y0, double y1,double dydx0, double dydx1,
-          bool computeIntegral, bool intx0x1, const std::string& name):
-_x0(x0),_x1(x1),_y0(y0),_y1(y1),_dydx0(dydx0),_dydx1(dydx1),
-     _computeIntegral(computeIntegral),_intx0x1(intx0x1),_name(name)
+SmoothSegmentedFunctionData::SmoothSegmentedFunctionData(
+    const SimTK::Matrix& mX,
+    const SimTK::Matrix& mY,
+    double x0,
+    double x1,
+    double y0,
+    double y1,
+    double dydx0,
+    double dydx1,
+    bool computeIntegral,
+    bool intx0x1,
+    const std::string& name) :
+    _x0(x0),
+    _x1(x1),
+    _y0(y0),
+    _y1(y1),
+    _dydx0(dydx0),
+    _dydx1(dydx1),
+    _computeIntegral(computeIntegral),
+    _intx0x1(intx0x1)
 {
-    
-
     _numBezierSections = mX.ncol();
 
     //////////////////////////////////////////////////
@@ -136,14 +459,14 @@ _x0(x0),_x1(x1),_y0(y0),_y1(y1),_dydx0(dydx0),_dydx1(dydx1),
             fitForSmoothingParameter(3,x,u,0).getSpline();
     }
 
-    if(_computeIntegral){
+    if(computeIntegral){
         //////////////////////////////////////////////////
         //Compute the integral of y(x) and spline the result    
         //////////////////////////////////////////////////
 
         SimTK::Matrix yInt =  SegmentedQuinticBezierToolkit::
             calcNumIntBezierYfcnX(xALL,0,INTTOL, UTOL, MAXITER,mX, mY,
-            _arraySplineUX,_intx0x1,_name);
+            _arraySplineUX,intx0x1,name);
 
         //not correct
         //if(_intx0x1==false){
@@ -163,55 +486,93 @@ _x0(x0),_x1(x1),_y0(y0),_y1(y1),_dydx0(dydx0),_dydx1(dydx1),
     }
 }
 
- SmoothSegmentedFunction::SmoothSegmentedFunction():
- _x0(SimTK::NaN),_x1(SimTK::NaN),_y0(SimTK::NaN)
-     ,_y1(SimTK::NaN),_dydx0(SimTK::NaN),_dydx1(SimTK::NaN),
-     _computeIntegral(false),_intx0x1(false),_name("NOT_YET_SET")
- {
-        _arraySplineUX.resize(0);        
-        _mXVec.resize(0);
-        _mYVec.resize(0);
-        _splineYintX = SimTK::Spline();
-        _numBezierSections = (int)SimTK::NaN;
-       
- }
+SmoothSegmentedFunction::SmoothSegmentedFunction(
+    const SimTK::Matrix& mX,
+    const SimTK::Matrix& mY,
+    double x0,
+    double x1,
+    double y0,
+    double y1,
+    double dydx0,
+    double dydx1,
+    bool computeIntegral,
+    bool intx0x1,
+    const std::string& name) :
+    _smoothData(
+        SmoothSegmentedFunctionDataLookup(
+            SmoothSegmentedFunctionParameters(
+                mX,
+                mY,
+                x0,
+                x1,
+                y0,
+                y1,
+                dydx0,
+                dydx1,
+                computeIntegral,
+                intx0x1),
+            name)
+        ),
+    _name(name)
+{}
 
- /*Detailed Computational Costs
- ________________________________________________________________________
-    If x is in the Bezier Curve
-                            Name     Comp.   Div.    Mult.   Add.    Assign.
+SmoothSegmentedFunction::SmoothSegmentedFunction() :
+    _smoothData(
+        SmoothSegmentedFunctionDataLookup(
+            SmoothSegmentedFunctionParameters(),
+            "NOT_YET_SET")
+        ),
+    _name("NOT_YET_SET")
+{}
+
+/*Detailed Computational Costs
+________________________________________________________________________
+   If x is in the Bezier Curve
+                           Name     Comp.   Div.    Mult.   Add.    Assign.
 _______________________________________________________________________
-        SegmentedQuinticBezierToolkit::
-                        calcIndex     3*m+2                   1*m     3   
-                            *calcU     15      2      82      42      60  
-        calcQuinticBezierCurveVal                     21      20      13
-                            total  15+3*m+2    2      103     62+1*m  76
+       SegmentedQuinticBezierToolkit::
+                       calcIndex     3*m+2                   1*m     3
+                           *calcU     15      2      82      42      60
+       calcQuinticBezierCurveVal                     21      20      13
+                           total  15+3*m+2    2      103     62+1*m  76
 
-        *Approximate. Uses iteration
+       *Approximate. Uses iteration
 ________________________________________________________________________
 If x is in the linear region
 
-                        Name     Comp.   Div.    Mult.   Add.    Assign.
-                                    1               1      2     1
+                       Name     Comp.   Div.    Mult.   Add.    Assign.
+                                   1               1      2     1
 ________________________________________________________________________
- 
- */
+
+*/
 
 double SmoothSegmentedFunction::calcValue(double x) const
 {
+    const SimTK::Array_<SimTK::Spline>& arraySplineUX =
+    _smoothData->_arraySplineUX;
+    const SimTK::Array_<SimTK::Vector>& mXVec = _smoothData->_mXVec;
+    const SimTK::Array_<SimTK::Vector>& mYVec = _smoothData->_mYVec;
+    double x0 = _smoothData->_x0;
+    double x1 = _smoothData->_x1;
+    double y0 = _smoothData->_y0;
+    double y1 = _smoothData->_y1;
+    double dydx0 = _smoothData->_dydx0;
+    double dydx1 = _smoothData->_dydx1;
+
     double yVal = 0;
-    if(x >= _x0 && x <= _x1 )
+
+    if(x >= x0 && x <= x1 )
     {
-        int idx  = SegmentedQuinticBezierToolkit::calcIndex(x,_mXVec);
+        int idx  = SegmentedQuinticBezierToolkit::calcIndex(x,mXVec);
         double u = SegmentedQuinticBezierToolkit::
-                 calcU(x,_mXVec[idx], _arraySplineUX[idx], UTOL,MAXITER);
+                 calcU(x,mXVec[idx], arraySplineUX[idx], UTOL,MAXITER);
         yVal = SegmentedQuinticBezierToolkit::
-                 calcQuinticBezierCurveVal(u,_mYVec[idx]);
+                 calcQuinticBezierCurveVal(u,mYVec[idx]);
     }else{
-        if(x < _x0){
-            yVal = _y0 + _dydx0*(x-_x0);            
+        if(x < x0){
+            yVal = y0 + dydx0*(x-x0);
         }else{
-            yVal = _y1 + _dydx1*(x-_x1);                    
+            yVal = y1 + dydx1*(x-x1);
         }    
     }
 
@@ -265,27 +626,35 @@ double SmoothSegmentedFunction::calcDerivative(double x, int order) const
 
     //QUINTIC SPLINE
 
-    
+    const SimTK::Array_<SimTK::Spline>& arraySplineUX =
+    _smoothData->_arraySplineUX;
+    const SimTK::Array_<SimTK::Vector>& mXVec = _smoothData->_mXVec;
+    const SimTK::Array_<SimTK::Vector>& mYVec = _smoothData->_mYVec;
+    double x0 = _smoothData->_x0;
+    double x1 = _smoothData->_x1;
+    double dydx0 = _smoothData->_dydx0;
+    double dydx1 = _smoothData->_dydx1;
+
     if(order==0){
                 yVal = calcValue(x);
     }else{
-            if(x >= _x0 && x <= _x1){        
-                int idx  = SegmentedQuinticBezierToolkit::calcIndex(x,_mXVec);
+            if(x >= x0 && x <= x1){
+                int idx  = SegmentedQuinticBezierToolkit::calcIndex(x,mXVec);
                 double u = SegmentedQuinticBezierToolkit::
-                                calcU(x,_mXVec[idx], _arraySplineUX[idx], 
+                                calcU(x,mXVec[idx], arraySplineUX[idx],
                                 UTOL,MAXITER);
                 yVal = SegmentedQuinticBezierToolkit::
-                            calcQuinticBezierCurveDerivDYDX(u, _mXVec[idx], 
-                            _mYVec[idx], order);
+                            calcQuinticBezierCurveDerivDYDX(u, mXVec[idx],
+                            mYVec[idx], order);
 /*
                             std::cout << _mX(3, idx) << std::endl;
                             std::cout << _mX(idx) << std::endl;*/
             }else{
                     if(order == 1){
-                        if(x < _x0){
-                            yVal = _dydx0;
+                        if(x < x0){
+                            yVal = dydx0;
                         }else{
-                            yVal = _dydx1;}
+                            yVal = dydx1;}
                     }else{
                         yVal = 0;}   
                 }
@@ -346,40 +715,49 @@ ________________________________________________________________________
 */
 double SmoothSegmentedFunction::calcIntegral(double x) const
 {
-    SimTK_ERRCHK1_ALWAYS(_computeIntegral,
+    SimTK_ERRCHK1_ALWAYS(_smoothData->_computeIntegral,
         "SmoothSegmentedFunction::calcIntegral",
         "%s: This curve was not constructed with its integral because"
         "computeIntegral was false",_name.c_str());
 
+    const SimTK::Spline& splineYintX = _smoothData->_splineYintX;
+    double x0 = _smoothData->_x0;
+    double x1 = _smoothData->_x1;
+    double y0 = _smoothData->_y0;
+    double y1 = _smoothData->_y1;
+    double dydx0 = _smoothData->_dydx0;
+    double dydx1 = _smoothData->_dydx1;
+    bool intx0x1 = _smoothData->_intx0x1;
+
     double yVal = 0;    
-    if(x >= _x0 && x <= _x1){
-        yVal = _splineYintX.calcValue(SimTK::Vector(1,x));
+    if(x >= x0 && x <= x1){
+        yVal = splineYintX.calcValue(SimTK::Vector(1,x));
     }else{
         //LINEAR EXTRAPOLATION         
-        if(x < _x0){
+        if(x < x0){
             SimTK::Vector tmp(1);
-            tmp(0) = _x0;
-            double ic = _splineYintX.calcValue(tmp);
-            if(_intx0x1){//Integrating left to right
-                yVal = _y0*(x-_x0) 
-                    + _dydx0*(x-_x0)*(x-_x0)*0.5 
+            tmp(0) = x0;
+            double ic = splineYintX.calcValue(tmp);
+            if(intx0x1){//Integrating left to right
+                yVal = y0*(x-x0)
+                    + dydx0*(x-x0)*0.5
                     + ic;
             }else{//Integrating right to left
-                yVal = -_y0*(x-_x0) 
-                    - _dydx0*(x-_x0)*(x-_x0)*0.5 
+                yVal = -y0*(x-x0)
+                    - dydx0*(x-x0)*(x-x0)*0.5
                     + ic;
             }            
         }else{
             SimTK::Vector tmp(1);
-            tmp(0) = _x1;
-            double ic = _splineYintX.calcValue(tmp);
-            if(_intx0x1){
-                yVal = _y1*(x-_x1) 
-                    + _dydx1*(x-_x1)*(x-_x1)*0.5 
+            tmp(0) = x1;
+            double ic = splineYintX.calcValue(tmp);
+            if(intx0x1){
+                yVal = y1*(x-x1)
+                    + dydx1*(x-x1)*(x-x1)*0.5
                     + ic;
             }else{
-                yVal = -_y1*(x-_x1) 
-                    - _dydx1*(x-_x1)*(x-_x1)*0.5 
+                yVal = -y1*(x-x1)
+                    - dydx1*(x-x1)*(x-x1)*0.5
                     + ic;
             }
         }
@@ -390,12 +768,12 @@ double SmoothSegmentedFunction::calcIntegral(double x) const
 
 bool SmoothSegmentedFunction::isIntegralAvailable() const
 {
-    return _computeIntegral;
+    return _smoothData->_computeIntegral;
 }
 
 bool SmoothSegmentedFunction::isIntegralComputedLeftToRight() const
 {
-    return _intx0x1;
+    return _smoothData->_intx0x1;
 }
 
 int SmoothSegmentedFunction::getArgumentSize() const
@@ -420,13 +798,15 @@ void SmoothSegmentedFunction::setName(std::string &name)
 
 SimTK::Vec2 SmoothSegmentedFunction::getCurveDomain() const
 {
+    const SimTK::Array_<SimTK::Vector>& mXVec = _smoothData->_mXVec;
+
     SimTK::Vec2 xrange;
     
     xrange(0) = 0; 
     xrange(1) = 0; 
-    if (!_mXVec.empty()) {
-        xrange(0) = _mXVec[0](0); 
-        xrange(1) = _mXVec[_mXVec.size()-1](_mXVec[0].size()-1); 
+    if (!mXVec.empty()) {
+        xrange(0) = mXVec[0](0);
+        xrange(1) = mXVec[mXVec.size()-1](mXVec[0].size()-1);
     }
     return xrange;
 }
@@ -473,26 +853,30 @@ SimTK::Matrix SmoothSegmentedFunction::calcSampledMuscleCurve(int maxOrder,
         "SmoothSegmentedFunction::calcSampledMuscleCurve",
         "Derivative order past the maximum computed order requested");
 
+    int numBezierSections = _smoothData->_numBezierSections;
+    bool computeIntegral = _smoothData->_computeIntegral;
+    const SimTK::Array_<SimTK::Vector>& mXVec = _smoothData->_mXVec;
+
     double x0,x1,delta;
     //y,dy,d1y,d2y,d3y,d4y,d5y,d6y,iy
-   SimTK::Vector midX(NUM_SAMPLE_PTS*_numBezierSections-(_numBezierSections-1));
+   SimTK::Vector midX(NUM_SAMPLE_PTS*numBezierSections-(numBezierSections-1));
    SimTK::Vector x(NUM_SAMPLE_PTS);
 
    //Generate a sample of X values inside of the curve that is denser where 
    //the curve is more curvy.
    double u;
    int idx = 0;
-      for(int s=0; s < _numBezierSections; s++){
+      for(int s=0; s < numBezierSections; s++){
         //Sample the local set for u and x
         for(int i=0;i<NUM_SAMPLE_PTS;i++){
                 u = ( (double)i )/( (double)(NUM_SAMPLE_PTS-1) );
                 x(i) = SegmentedQuinticBezierToolkit::
-                    calcQuinticBezierCurveVal(u,_mXVec[s]);    
-                if(_numBezierSections > 1){
+                    calcQuinticBezierCurveVal(u,mXVec[s]);
+                if(numBezierSections > 1){
                    //Skip the last point of a set that has another set of points
                    //after it. Why? The last point and the starting point of the
                    //next set are identical in value.
-                    if(i<(NUM_SAMPLE_PTS-1) || s == (_numBezierSections-1)){
+                    if(i<(NUM_SAMPLE_PTS-1) || s == (numBezierSections-1)){
                         midX(idx) = x(i);
                         idx++;
                     }
@@ -508,7 +892,7 @@ SimTK::Matrix SmoothSegmentedFunction::calcSampledMuscleCurve(int maxOrder,
     
     SimTK::Matrix results;
 
-    if(_computeIntegral){
+    if(computeIntegral){
         results.resize(pts*(midX.size()-1)+2*10*pts,maxOrder+2+1);
     }else{
         results.resize(pts*(midX.size()-1)+2*10*pts,maxOrder+2);
@@ -529,11 +913,11 @@ SimTK::Matrix SmoothSegmentedFunction::calcSampledMuscleCurve(int maxOrder,
 
     //generate some sample points in the extrapolated region
     idx=0;
-    x0 = _x0 - 0.1*(_x1-_x0);
+    x0 = _smoothData->_x0 - 0.1*(_smoothData->_x1-_smoothData->_x0);
     if(domainMin < x0)
         x0 = domainMin;
 
-    x1 = _x0;
+    x1 = _smoothData->_x0;
     delta = (0.1)*(x1-x0)/(pts);
 
     for(int j=0;j<pts*10;j++){
@@ -554,8 +938,8 @@ SimTK::Matrix SmoothSegmentedFunction::calcSampledMuscleCurve(int maxOrder,
     }
 
     //generate some points in the extrapolated region
-    x0 = _x1;
-    x1 = _x1 + 0.1*(_x1-_x0);
+    x0 = _smoothData->_x1;
+    x1 = _smoothData->_x1 + 0.1*(_smoothData->_x1-_smoothData->_x0);
     if(domainMax > x1)
         x1 = domainMax;
 
@@ -590,7 +974,7 @@ SimTK::Matrix SmoothSegmentedFunction::calcSampledMuscleCurve(int maxOrder,
         if(maxOrder>=6)
         results(i,7) = calcDerivative(d6y,ax);
 
-        if(_computeIntegral){
+        if(computeIntegral){
             results(i,maxOrder+2) = calcIntegral(ax(0));
         }
     }